Add delete list item functionality. Fix issue with capitalization discrepancies.

+++ app.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 // const date = require(__dirname + "/date.js");
 
 const app = express();
@@ -74,7 +75,7 @@ app.get("/", (req, res) => {
 
 // CUSTOM LIST ROUTE //
 app.get("/:customListName", (req, res) => {
-    const customListName = req.params.customListName;
+    const customListName = _.capitalize(req.params.customListName);
 
     List.findOne({name: customListName}, (err, foundList) => {
         if (err) {
@@ -155,14 +156,28 @@ app.post("/", (req, res) => {
 /////////////////
 app.post("/delete", (req, res) => {
     const checkedItemId = req.body.checkbox;
+    const listName = req.body.listName;
 
-    Item.findByIdAndRemove(checkedItemId, (err) => {
-        if (err) {
-            console.log(err);
-        } else {
-            res.redirect("/");
-        }
-    });
+    if (listName === "Today") {
+        Item.findByIdAndRemove(checkedItemId, (err) => {
+            if (err) {
+                console.log(err);
+            } else {
+                res.redirect("/");
+            }
+        });
+    } else {
+        // find the document we want to update and remove the item from our list
+        List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}}, (err, foundList) => {
+            if (!err) {
+                res.redirect("/" + listName);
+            } else {
+                console.log(err);
+            }
+        });
+    }
+
+
 });
